Move the error_codes.json into the library and change the test cases
