+++ test/test_attribute_exists.js
@@ -5,7 +5,7 @@ const should = require('should');
 const config = require('./config.json');
 const StateError = require('../libs/errors/state_error');
 const errorHandler = require('../libs/errors/error_dispenser').errorFunction;
-const errorCodes = require('./error_codes');
+const errorCodes = require('../libs/error_codes');
 const errorMessages = require('../libs/messages.json');
 
 describe('Testing the attributeExists function', () => {
