+++ test/test_compare.js
@@ -5,7 +5,7 @@ const should = require('should');
 const config = require('./config.json');
 const StateError = require('../libs/errors/state_error');
 const errorHandler = require('../libs/errors/error_dispenser').errorFunction;
-const errorCodes = require('./error_codes');
+const errorCodes = require('../libs/error_codes');
 const errorMessages = require('../libs/messages.json');
 
 describe('Testing the Compare function', () => {
@@ -34,10 +34,10 @@ describe('Testing the Compare function', () => {
     return ldapAsyncWrap.compare(1, attr, val)
       .then(() => { should.fail('should not have succeeded'); })
       .catch(
-      TypeError,
-      (error) => {
-        should.deepEqual(error.message, errorMessages.typeErrorMessage);
-      })
+        TypeError,
+        (error) => {
+          should.deepEqual(error.message, errorMessages.typeErrorMessage);
+        })
       .catch((err) => { should.fail('did not expect generic error'); });
   });
 
@@ -54,11 +54,11 @@ describe('Testing the Compare function', () => {
     return ldapAsyncWrap.compare(dn, noSuchAttribute, val)
       .then(() => { should.fail('should not have passed'); })
       .catch(
-      CustomError,
-      (err) => {
-        should.deepEqual(
-          err.constructor.description, CustomError.description);
-      })
+        CustomError,
+        (err) => {
+          should.deepEqual(
+            err.constructor.description, CustomError.description);
+        })
       .catch((err) => { should.fail('did not expect generic error'); });
   });
 
@@ -77,11 +77,11 @@ describe('Testing the Compare function', () => {
     return ldapAsyncWrap.compare(dn, nonAttr, val)
       .then(() => { should.fail('should not have succeeded'); })
       .catch(
-      CustomError,
-      (err) => {
-        should.deepEqual(
-          err, new CustomError(errorMessages.ldapCompareErrorMessage));
-      })
+        CustomError,
+        (err) => {
+          should.deepEqual(
+            err, new CustomError(errorMessages.ldapCompareErrorMessage));
+        })
       .catch((err) => { should.fail('did not expect generic error'); });
   });
 
@@ -92,11 +92,11 @@ describe('Testing the Compare function', () => {
     return ldapAsyncWrap.compare(nonObj, attr, val)
       .then(() => { should.fail('should not have succeeded'); })
       .catch(
-      CustomError,
-      (err) => {
-        should.deepEqual(
-          err, new CustomError(errorMessages.ldapCompareErrorMessage));
-      })
+        CustomError,
+        (err) => {
+          should.deepEqual(
+            err, new CustomError(errorMessages.ldapCompareErrorMessage));
+        })
       .catch((err) => { should.fail('did not expect generic error'); });
   });
 
@@ -114,11 +114,11 @@ describe('Testing the Compare function', () => {
       .then(() => { return ldapAsyncWrap.compare(dn, attr, val); })
       .then(() => { should.fail('should not have succeeded'); })
       .catch(
-      CustomError,
-      (err) => {
-        should.deepEqual(
-          err, new CustomError(errorMessages.ldapCompareErrorMessage));
-      })
+        CustomError,
+        (err) => {
+          should.deepEqual(
+            err, new CustomError(errorMessages.ldapCompareErrorMessage));
+        })
       .catch((err) => { should.fail('did not expect generic error'); });
   });
 
@@ -134,10 +134,10 @@ describe('Testing the Compare function', () => {
       .catch(() => { return ldapAsyncWrap.compare(dn, attr, val); })
       .then(() => { should.fail('should not have succeeded'); })
       .catch(
-      StateError,
-      (err) => {
-        should.deepEqual(err.message, errorMessages.bindErrorMessage);
-      })
+        StateError,
+        (err) => {
+          should.deepEqual(err.message, errorMessages.bindErrorMessage);
+        })
       .catch((err) => { should.fail('did not expect generic error'); });
   });
 
@@ -149,10 +149,10 @@ describe('Testing the Compare function', () => {
         .then(() => { return ldapAsyncWrap.compare(dn, attr, val); })
         .then(() => { should.fail('should not have succeeded'); })
         .catch(
-        StateError,
-        (err) => {
-          should.deepEqual(err.message, errorMessages.bindErrorMessage);
-        })
+          StateError,
+          (err) => {
+            should.deepEqual(err.message, errorMessages.bindErrorMessage);
+          })
         .catch((err) => { should.fail('did not expect generic error'); });
     });
 
@@ -162,10 +162,10 @@ describe('Testing the Compare function', () => {
       .then(() => { return ldapAsyncWrap.compare(dn, attr, val); })
       .then(() => { should.fail('should not have succeeded'); })
       .catch(
-      StateError,
-      (err) => {
-        should.deepEqual(err.message, errorMessages.bindErrorMessage);
-      })
+        StateError,
+        (err) => {
+          should.deepEqual(err.message, errorMessages.bindErrorMessage);
+        })
       .catch((err) => { should.fail('did not expect generic error'); });
   });
 
@@ -196,12 +196,12 @@ describe('Testing the Compare function', () => {
         })
         .then(() => { should.fail('should not have succeeded'); })
         .catch(
-        CustomError,
-        (err) => {
-          should.deepEqual(
-            err, new CustomError(errorMessages.ldapCompareErrorMessage));
-          return ldapAsyncWrap.compare(dn, attr, nonVal);
-        })
+          CustomError,
+          (err) => {
+            should.deepEqual(
+              err, new CustomError(errorMessages.ldapCompareErrorMessage));
+            return ldapAsyncWrap.compare(dn, attr, nonVal);
+          })
         .catch(() => { should.fail('did not expect generic error'); })
         .then((result3) => { should.deepEqual(result3, false); });
     });
