+++ test/test_delete.js
@@ -6,7 +6,7 @@ const Promise = require('bluebird');
 const config = require('./config.json');
 const errorHandler = require('../libs/errors/error_dispenser').errorFunction;
 const StateError = require('../libs/errors/state_error');
-const errorCodes = require('./error_codes');
+const errorCodes = require('../libs/error_codes');
 const errorMessages = require('../libs/messages.json');
 
 describe('Testing the async LDAP delete operation', () => {
