+++ test/test_modify.js
@@ -7,7 +7,7 @@ const Promise = require('bluebird');
 const errorHandler = require('../index').errorHandler;
 const StateError = require('../libs/errors/state_error');
 const ValidationError = require('../libs/errors/validation_error');
-const errorCodes = require('./error_codes');
+const errorCodes = require('../libs/error_codes');
 const errorMessages = require('../libs/messages.json');
 
 describe('Testing the modify functionalities', () => {
@@ -88,8 +88,8 @@ describe('Testing the modify functionalities', () => {
   it('should replace the old attributes with new one from an entry', () => {
     return ldapAsyncWrap
       .modify(
-      config.ldapModify.ldapModificationReplace.change_dn,
-      changeAttributesReplace)
+        config.ldapModify.ldapModificationReplace.change_dn,
+        changeAttributesReplace)
       .then((result) => { should.deepEqual(result, 0); });
   });
 
@@ -128,7 +128,7 @@ describe('Testing the modify functionalities', () => {
 
     return ldapAsyncWrap
       .modify(
-      config.ldapModify.ldapModificationReplace.change_dn, attribute)
+        config.ldapModify.ldapModificationReplace.change_dn, attribute)
       .then(() => {
         should.fail('should not have succeeded');
       })
@@ -149,8 +149,8 @@ describe('Testing the modify functionalities', () => {
     };
     return ldapAsyncWrap
       .modify(
-      config.ldapModify.ldapModificationReplace.change_dn,
-      changeAttributes, control)
+        config.ldapModify.ldapModificationReplace.change_dn,
+        changeAttributes, control)
       .then((result) => {
         let resultOperation;
         resultOperation = result.split('\n');
@@ -168,8 +168,8 @@ describe('Testing the modify functionalities', () => {
     }];
     return ldapAsyncWrap
       .modify(
-      config.ldapModify.ldapModificationReplace.change_dn,
-      changeAttributes, control)
+        config.ldapModify.ldapModificationReplace.change_dn,
+        changeAttributes, control)
       .then(() => {
         should.fail('should not have succeeded');
       })
@@ -199,8 +199,8 @@ describe('Testing the modify functionalities', () => {
   it('should add new attributes to an existing entry', () => {
     return ldapAsyncWrap
       .modify(
-      config.ldapModify.ldapModificationReplace.change_dn,
-      changeAttributesAdd)
+        config.ldapModify.ldapModificationReplace.change_dn,
+        changeAttributesAdd)
       .then((result) => { should.deepEqual(result, 0); });
   });
 
@@ -215,8 +215,8 @@ describe('Testing the modify functionalities', () => {
     ];
     return ldapAsyncWrap
       .modify(
-      config.ldapModify.ldapModificationReplace.change_dn,
-      changeAttributesAdd)
+        config.ldapModify.ldapModificationReplace.change_dn,
+        changeAttributesAdd)
       .then((result) => { should.fail('should not have passed'); })
       .catch(CustomError, (err) => {
         should.deepEqual(err.constructor.description, CustomError.description);
@@ -237,8 +237,8 @@ describe('Testing the modify functionalities', () => {
     ];
     return ldapAsyncWrap
       .modify(
-      config.ldapModify.ldapModificationReplace.change_dn,
-      change)
+        config.ldapModify.ldapModificationReplace.change_dn,
+        change)
       .then((result) => { should.fail('should not have passed'); })
       .catch(CustomError, (err) => {
         should.deepEqual(err.constructor.description, CustomError.description);
@@ -259,8 +259,8 @@ describe('Testing the modify functionalities', () => {
     ];
     return ldapAsyncWrap
       .modify(
-      config.ldapModify.ldapModificationReplace.change_dn,
-      change)
+        config.ldapModify.ldapModificationReplace.change_dn,
+        change)
       .then((result) => { should.fail('should not have passed'); })
       .catch(CustomError, (err) => {
         should.deepEqual(err.constructor.description, CustomError.description);
@@ -274,16 +274,16 @@ describe('Testing the modify functionalities', () => {
   it('should delete an attribute from an existing entry', () => {
     return ldapAsyncWrap
       .modify(
-      config.ldapModify.ldapModificationReplace.change_dn,
-      changeAttributesDelete)
+        config.ldapModify.ldapModificationReplace.change_dn,
+        changeAttributesDelete)
       .then((result) => { should.deepEqual(result, 0); });
   });
 
   it('should make multiple modifications to an entry', () => {
     return ldapAsyncWrap
       .modify(
-      config.ldapModify.ldapModificationReplace.change_dn,
-      changeAttributes)
+        config.ldapModify.ldapModificationReplace.change_dn,
+        changeAttributes)
       .then((result) => { should.deepEqual(result, 0); });
   });
 
@@ -328,8 +328,8 @@ describe('Testing the modify functionalities', () => {
   it('should return a specific attribute from the entry', () => {
     return ldapAsyncWrap
       .modify(
-      config.ldapModify.ldapModificationReplace.change_dn,
-      changeAttributes, controlOperation)
+        config.ldapModify.ldapModificationReplace.change_dn,
+        changeAttributes, controlOperation)
       .then((result) => {
         let resultOperation;
         resultOperation = result.split('\n');
